[WIP] Backport module from 10.0 to 9.0
**Not installable, possible critical issue to backport:** https://github.com/odoo/odoo/issues/16540

In order to make this installable and functional, we need to replace the reference to `_inherit = 'base'` (in the `queue_job/models/base.py` file) but the `name = 'base'` reference in core (`models.py`) was not added until version 10.0. There are no apparent plans to backport that change and no obvious workarounds.